code/requirements.txt
numpy==2.1.3
pandas==2.2.3
geopandas==1.0.1
astropy==6.1.5
scipy==1.14.1
pillow==9.0.1
tqdm==4.67.0
matplotlib==3.9.2
pyspeckit
dash-leaflet
dash-extensions